Change default installed project version used in benchmark discovery
Choose either a most recently built version from the wheel cache, or if
none found, use the current head of the master branch.

This speeds up benchmark discovery avoiding rebuilding when possible,
and avoids build failures that occur if the project repo was previously
at an unbuildable commit. Instead, we assume now that the master branch
is at an installable state.