Restructure repository management logic to act as a mirror of remote repository
Make the Repo object act transparently as a mirror of the remote
repository.  Also rework the logic so that a checkout of the working
tree is done only in the environment-specific clones.

This makes remote branches available for all commands without special
prefixes, and removes the need for Repo.check_remote_branch.

Also rename some of the commands, so that branch/tag/commit is called
"name", and replace checkout_parent with get_hash_from_parent.