BUG: do not rerun setup() when determining `number` in timing benchmarks
Do not re-run setup in the loop that determines what number of calls to
make without repeats.  This can be important in cases where setup()
takes a lot of time, but the timing routine itself if fast.

Also fix a logic bug for cases where the loop runs all iterations; in
this case, the `timing` corresponded to `number//10` resulting to wrong
reported timing for the first repeat.