BUG: fix two issues with asv publish
The first issue occurred if a JSON file contained an integer and another contained a string, for the same parameter (for instance, for RAM, a file might contain “8G” and another file 104727291).

The second issue occurred if one file contained an empty string and another contained None, for the same parameter, since this would result in two empty strings in the final list and would then create too many entries in the generated website.