mercurial: Only follow first parent on merges
Like in the git plugin we follow only the first parent on merges with
the followfirst option.

To make followfirst work properly on any revset we must ensure the
revset is sorted by revision desc.

In the case of the common workflow with two branches (default and
stable) and merge between them, this fix the wrong behavior of the
BranchCache reporting that most of changesets are on all branches.