Handle benchmarks by commit revision instead of date
Commits date are not necessary monotonic along the commit log,
especially if working with rebase workflow (instead of merge), this
generate false positive or missed regression detection and unwanted
peaks in the web UI.

This patch introduce a revision number as a drop-in replacement of
commit date. This revision number is an incremental number representing
the number of ancestors commits.

In the web UI the distance between commits is the number of commits
between them. Commits can still be displayed by date using the date
selector.

Reverting cc6accae since revisions are no longer long ints and add a
test on summary graph resampling.

Closes #390