Implement benchmark result versioning
Add a version attribute to benchmarks, and when benchmarks are run,
tag the results produced with it. The default value is based on hash of
source code lines, but can be overridden by user.

In run, consider results with different benchmark version as missing.
In compare, mark incompatible results.
In publish, ignore results with different benchmark versions.