chmod html_dir to ensure writability (#521)
asv's HTML output is made by copying a "template" directory, which lives
alongside asv's code (e.g. in /usr or the like), to the destination, which is
then modified to contain the benchmark data.

If the template is read-only, e.g. due to the filesystem it's stored on, then we
can copy it across, but don't have permission to modify it.