Perform benchmark discovery only when running benchmarks
Discover benchmarks only when the project is going to be built and the
commit hashes to be checked are known, so the benchmark discovery can be
performed using those commits.  This also adjusts the Benchmarks API so
that the discovery is explicit.

Previously, benchmark discovery was done installing a version of the
project at the tip of the branch or from an existing wheel. This leads
to problems when the build fails or the project is not in an importable
state.

This also changes "asv publish" to emit an error message rather than
running benchmark discovery, if benchmarks.json is damaged or missing.
That is an unusual error condition, so it is reasonable to handle it as
such (the error message explains how to resolve it).