Fix benchmark name displayed in summarygrid
Commit c72518a (Display last part of benchmark name in left pannel)
wasn't working before when no pretty_name is set for the benchmark since
it was defaulting to the full benchmark name.

I see no reason to have a default for pretty_name and it's better to
have pretty_name undefined in javascript in this case, so let's drop the
default.