Fix #547: allow specifying project subdir in the repo
Some Python projects are not located at the root of their repository.
For example, the Arrow project has its Python bindings in a `python`
subdirectory (see https://github.com/apache/arrow/).

This allows configuring ASV so that it is able to run "setup.py build"
and "pip install" from the right directory.