Add a min_repeat benchmark attribute
When sample_time is lower than a single benchmark we run only one
sample. For some benchmark it's hard to set sample time to a proper
value to have at least a sufficient number of samples (depending on
hardware / parameterization etc).
Add a min_repeat benchmark attribute to ensure we return a value
computed on at least a given number of samples.

Closes #653