benchmark: revise timing benchmark 'repeat' determination + adjust 'sample_time'
Change default for sample_time to 10ms. This is sufficiently long for
obtaining accurate measurements with most timers, with reasonable
accuracy retained.  Since asv runs a large number of benchmarks, it is
important that the measurement times are kept short.

Make 'repeat' determination to work on the basis of a given range
(min_repeat, max_repeat) and a maximum allocated time max_time.

Change min_timeit_count to min_run_count. This is retained separately
from min_repeat, as its purpose is different (e.g. deal with numba jit
compilation) and it is counted also during warmup.