Use 'docutils' instead of 'six' as an installable package in tests
It's not possible to replace 'six' with a dummy package, as some
versions of wheel depend on it.

Ideally, we'd use dummy packages not existing in Pypi, but at the moment
this is limited by the need to have the same packages available in conda
(and we don't build dummy conda packages yet).