benchmark: use pkgutil to recursively iterate over all modules
Use a more standard way (pkgutil) to discover all modules in the
benchmark package.

This is a slight change in the API: previously any '.py' files in the
bencmarks directory were discovered, but with this change they are only
discovered insofar as they are in a valid Python package hierarchy.