detect_regression: value 'after' regression is the minimum, not next one
Change the value 'after' a regression to be the minimum of values
following the upward step.

Previously, it was the value immediately after the upward step.
However, typically the important question is how much worse the
situation is, after taking into account any commits trying to rectify
the problem.

Using 'minimum' instead of 'next' value enables better ranking of how
important regressions are in practice.

That the code becomes also simpler is probably a sign that this is the
right thing.