Rename load_json cleanup= option to js_comments=
The JSON comment syntax is nonstandard, and significantly reduces the
loading speed.

Rename the option and change the default, to ensure is is not used
unnecessarily.

Also remove Config.update --- since update_json throws away the
comments, it's not really possible to update the API number in the
config file.