Change suggested badge colour
I suggest to change the suggested badge-color from green to blue. Red-green is often used to show if tests or builds were successful or not (readthedocs, travis, etc), and red-orange-green is often used to show the status (coveralls, codacy). However, the asv-badge is a static link, indicating that it is benchmarked with asv. So I suggest to show it in blue, as for instance used by Zenodo for its doi, or py pypi and conda for the version number; static info.