ENH: add --date-period option to asv run for 1 commit / n days
Add a new filtering option --date-period to asv run, for subsampling the
specified range of commits so that the selected commits are spaced in
time by the given period.

The initial part of the selection remains stable when new commits are
added, so this is suitable also for cron jobs.