Add --strict flag to exit non-zero on benchmark failures
This commit adds a new flag, --strict, to the run (and by extension dev)
and continous commands. When set any benchmark failures encountered
will be treated as an error and the command will exit non-zero when the
run is finished. This is especially useful when running benchmarks in
a CI system or similar automated environment and wish to know if a
change was made to break a benchmark.

Fixes #449