Adjust failure return code and tracking based on review
This commit fixes issues identified by review. Mainly that trying to set
the return code to be the number of failures is prone to be unreliable
because if it exceeds the max size for a response code. To address that
this uses 2 as the return code now if any failures are found.
Additionally, the failures tracking list no longer needs to be a list if
we are not counting the number of failures to set the return code
anymore. This changes it to be a boolean and simplifies the logic to
just track whether we've found any errors.