MAINT: add support for running tox (#867)
This commit adds a tox.ini file to support using tox to run unit tests
locally. Tox makes it simple for users and contributors to run tests
locally with just the tox command. It will handle building a venv
installing asv in that venv and then running tests. It's especially
useful for testing multiple versions of python since it will build and
maintain separate testing venvs for running with multiple python
interpreters.