Fix JavaScript navigation when method names contain escaped HTML characters
In Python 3, a method such as `def time_β(): pass` is valid. However,
asv's HTML output struggles with characters in benchmark names that get
URL encoded. The following shows up in the browser console for such
a method:

```
Uncaught TypeError: index.benchmarks[current_benchmark] is undefined
```

The problem here seems to be that `current_benchmark` is not correctly
decoded and contains the URL encoding of `β`.