Fix pip requirement building in new environments
By default it seems that:

- pip+build
- pip+asv_runner
- pip+pympler

make it to the requirements of building your conda environment

Without this PR, the command gets executed and adds things like like a
bogus environment.yml (temporary only) that looks lik

```
name: conda-py3.12
channels:
   - conda-forge
```

but this causes conda to error since there are no requirements for the
"update".

We can debate as to whether or not these "pip dependencies" should be
converted to conda ones, but lets leave that debate for later. This
helps people run "asv setup"