feat: Add verbose logging when we fail, not subcommands
Fixes: #1453

The combination of build and install commands can fail in many
ways.  When users enable verbose asv, we should let them know
if the failure occurs in our logic.

In this particular case, it was due to the build command placing
multiple wheels inside the build_cache_dir.  The failure occured
because we assume exactly one wheel is produced in
build_cache_dir, not the wheels for all dependencies.

Without verbose logging, the only way to identify the bad
build/install command interaction is to debug.